requirements.txt
langchain
langchain-groq
python-dotenv
langchain-openai
openai